Reorganize layout components and css
