+++ src/css/page-transition-loading-animation.css
